+++ dlkp/models/ke/kpe.py
@@ -45,181 +45,27 @@ from transformers import (
 )
 from transformers.trainer_utils import get_last_checkpoint, is_main_process
 
+from transformer.crf_models import BERT_CRFforTokenClassification
+from transformer.token_classification_models import LongformerForTokenClassification
+from crf.crf_trainer import CRF_Trainer
+from extraction_utils import ModelArguments, DataTrainingArguments
 
 logger = logging.getLogger(__name__)
-# from models.long_doc_kp_models import LONG_DOC_KP_MODELS
 
-# KPE_MODELS_DICT={
-#     'others': AutoModelForTokenClassification,
-#     "longformer":
-#     'reformer' :
-#     'crf_longformer' :
-#     'crf_bert': BERT_CRFforTokenClassification
-# }
 
 CRF_MODEL_DICT = {
     "bert": BERT_CRFforTokenClassification,
-    "longformer": Longformer_CRFforTokenClassification,
+    # "longformer": Longformer_CRFforTokenClassification,
 }
 TOKEN_MODEL_DICT = {
     "bert": BertForTokenClassification,
-    "longformer": LongformerForTokenClassification,
-    "reformer": ReformerForTokenClassification,
+    # "longformer": LongformerForTokenClassification,
+    # "reformer": ReformerForTokenClassification,
 }
 
 MODEL_DICT = {"crf": CRF_MODEL_DICT, "simple": TOKEN_MODEL_DICT}
 
-# KPE_MODELS_DICT = KPE_MODELS_DICT | LONG_DOC_KP_MODELS
 
-
-@dataclass
-class ModelArguments:
-    """
-    Arguments pertaining to which model/config/tokenizer we are going to fine-tune from.
-    """
-
-    model_family_name: str = field(
-        metadata={
-            "help": "name of the family of model, bert, longformer, reformer etc."
-        }
-    )
-    model_name_or_path: str = field(
-        metadata={
-            "help": "Path to pretrained model or model identifier from huggingface.co/models"
-        }
-    )
-    config_name: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "Pretrained config name or path if not the same as model_name"
-        },
-    )
-    tokenizer_name: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "Pretrained tokenizer name or path if not the same as model_name"
-        },
-    )
-    cache_dir: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "Where do you want to store the pretrained models downloaded from huggingface.co"
-        },
-    )
-    model_revision: str = field(
-        default="main",
-        metadata={
-            "help": "The specific model version to use (can be a branch name, tag name or commit id)."
-        },
-    )
-    use_CRF: bool = field(
-        default=False,
-        metadata={"help": "wether to use CRF on top of the classifier"},
-    )
-    use_BiLSTM: bool = field(
-        default=False,
-        metadata={"help": "use BiLSTM in sequence classification"},
-    )
-
-
-@dataclass
-class DataTrainingArguments:
-    """
-    Arguments pertaining to what data we are going to input our model for training and eval.
-    """
-
-    task_name: Optional[str] = field(
-        default="simple", metadata={"help": "The name of the task simple, crf"}
-    )
-
-    train_file: Optional[str] = field(
-        default=None,
-        metadata={"help": "The input training data file (a csv or JSON file)."},
-    )
-    validation_file: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "An optional input evaluation data file to evaluate on (a csv or JSON file)."
-        },
-    )
-    test_file: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "An optional input test data file to predict on (a csv or JSON file)."
-        },
-    )
-    overwrite_cache: bool = field(
-        default=False,
-        metadata={"help": "Overwrite the cached training and evaluation sets"},
-    )
-    preprocessing_num_workers: Optional[int] = field(
-        default=None,
-        metadata={"help": "The number of processes to use for the preprocessing."},
-    )
-    pad_to_max_length: bool = field(
-        default=False,
-        metadata={
-            "help": "Whether to pad all samples to model maximum sentence length. "
-            "If False, will pad the samples dynamically when batching to the maximum length in the batch. More "
-            "efficient on GPU but very bad for TPU."
-        },
-    )
-    label_all_tokens: bool = field(
-        default=False,
-        metadata={
-            "help": "Whether to put the label for one word on all tokens of generated by that word or just on the "
-            "one (in which case the other tokens will have a padding index)."
-        },
-    )
-    return_entity_level_metrics: bool = field(
-        default=False,
-        metadata={
-            "help": "Whether to return all the entity levels during evaluation or just the overall ones."
-        },
-    )
-    dataset_name: Optional[str] = field(
-        default=None,
-        metadata={"help": "The name of the dataset to use (via the datasets library)."},
-    )
-    dataset_config_name: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "The configuration name of the dataset to use (via the datasets library)."
-        },
-    )
-    cache_file_name: Optional[str] = field(
-        default=None,
-        metadata={
-            "help": "Provide the name of a path for the cache file. It is used to store the results of the computation instead of the automatically generated cache file name."
-        },
-    )
-
-    def __post_init__(self):
-        if (
-            self.dataset_name is None
-            and self.train_file is None
-            and self.validation_file is None
-        ):
-            raise ValueError(
-                "Need either a dataset name or a training/validation file."
-            )
-        else:
-            if self.train_file is not None:
-                extension = self.train_file.split(".")[-1]
-                assert extension in [
-                    "csv",
-                    "json",
-                ], "`train_file` should be a csv or a json file."
-            if self.validation_file is not None:
-                extension = self.validation_file.split(".")[-1]
-                assert extension in [
-                    "csv",
-                    "json",
-                ], "`validation_file` should be a csv or a json file."
-        self.task_name = self.task_name.lower()
-
-
-# def main():
 TRAINER_DICT = {
     "crf": CRF_Trainer,
     "simple": Trainer,
